getIconUrl error

+++ lib/models/weather_seven_days_daily_model.dart
@@ -217,8 +217,8 @@ class FeelsLike {
   FeelsLike.fromJson(Map<String, dynamic> json) {
     day = json['day'].toDouble();
     night = json['night'].toDouble();
-    eve = json['eve'];
-    morn = json['morn'];
+    eve = json['eve'].toDouble();
+    morn = json['morn'].toDouble();
   }
 
   Map<String, dynamic> toJson() {
